# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  partner.ReferralLinkLookup:
    properties:
      code:
        type: string
      isActive:
        type: boolean
      partnerEmail:
        type: string
      partnerFullName:
        type: string
      partnerId:
        type: string
    type: object
...
  structs.ErrorResponse:
    properties:
      body:
        type: string
      error_msg:
        type: string
      status_code:
        type: integer
    type: object
...
  structs.ResponseBody:
    properties:
      body: {}
      message:
        type: string
    type: object
...
  structs.SuccessResponse:
    properties:
      success:
        type: boolean
    type: object
...
  structs.TradeEventParams:
    properties:
      symbol:
        type: string
      tradeAmount:
        type: number
      tradeFee:
        type: number
      tradeId:
        type: string
      tradeTimestamp:
        type: integer
      userId:
        type: string
    required:
    - tradeAmount
    - tradeFee
    - tradeId
    - userId
    type: object
  structs.UnlinkReferralEventParams:
    properties:
      userId:
        type: string
    required:
    - userId
    type: object
  structs.UserDepositedParams:
    properties:
      amount:
        type: number
      userId:
        type: string
    required:
    - amount
    - userId
    type: object
  structs.UserRegisteredParams:
    properties:
      referralCode:
        type: string
      userId:
        type: string
    required:
    - referralCode
    - userId
    type: object
host: localhost:8090
info:
  contact: {}
  description: |-
    Partner program backend — public, partner, admin, system endpoints.
    Authentication is split into three layers:
    • BearerAuth (PartnerAuth/AdminAuth): Cognito ID token in `Authorization: Bearer …`.
    • InternalKey: `X-Internal-Key` header — only the xmeta-monorepo server should send this.
  title: xmeta Partner API
  version: "1.0"
paths:
...
  /internal/link-referral:
    post:
      consumes:
      - application/json
      description: Called by xmeta-monorepo right after Cognito signup completes for
        any user that arrived through a `?ref=CODE` link. Mirrors `POST /partner/auth/link-referral`
        but is server-to-server (X-Internal-Key) instead of user-authenticated.
      parameters:
      - description: Referral attach payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.UserRegisteredParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/structs.ResponseBody'
            - properties:
                body:
                  $ref: '#/definitions/structs.SuccessResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - InternalKey: []
      summary: Link a freshly signed-up user to a partner
      tags:
      - System Events
  /internal/referral-links/{code}:
    get:
      description: Looks up a referral code so the monorepo signup/settings form can
        show real-time "valid / belongs to {Name}" feedback before the user submits.
        Returns 404 if the code is unknown.
      parameters:
      - description: Referral code, e.g. ABC1234
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/structs.ResponseBody'
            - properties:
                body:
                  $ref: '#/definitions/partner.ReferralLinkLookup'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - InternalKey: []
      summary: Validate a referral code
      tags:
      - System Events
  /internal/trade-event:
    post:
      consumes:
      - application/json
      description: Ingests a trade event from monorepo and runs the commission engine
      parameters:
      - description: Trade event payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.TradeEventParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/structs.ResponseBody'
            - properties:
                body:
                  $ref: '#/definitions/structs.SuccessResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - InternalKey: []
      summary: Process a trade event
      tags:
      - System Events
  /internal/unlink-referral:
    post:
      consumes:
      - application/json
      description: Server-side counterpart to `POST /partner/auth/unlink-referral`.
        Called by xmeta-monorepo on account closure, compliance flags, or other system-driven
        detachments. Past commissions stay attributed to whoever was active at trade
        time.
      parameters:
      - description: Unlink payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.UnlinkReferralEventParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/structs.ResponseBody'
            - properties:
                body:
                  $ref: '#/definitions/structs.SuccessResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - InternalKey: []
      summary: Unlink a user from their current partner
      tags:
      - System Events
  /internal/user-deposited:
    post:
      consumes:
      - application/json
      description: Records a referred user deposit and runs deposit-based commission
        rules
      parameters:
      - description: User deposit payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.UserDepositedParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/structs.ResponseBody'
            - properties:
                body:
                  $ref: '#/definitions/structs.SuccessResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - InternalKey: []
      summary: Process user deposit
      tags:
      - System Events
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: '"Bearer {Cognito ID Token}" format'
    in: header
    name: Authorization
    type: apiKey
  InternalKey:
    description: Server-to-server API key shared with xmeta-monorepo
    in: header
    name: X-Internal-Key
    type: apiKey
swagger: "2.0"